backend/requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
groq
python-dotenv
httpx
